Added a /configupdate command (that doesn't work, sue me).

+++ handlers/ConfigHandler.js
@@ -8,20 +8,20 @@ const ConfigErrorHandler = require("./ConfigErrorHandler");
  * 
  * TODO:
  * - Allow for the creation of the file (with pre-defined defaults) 
- *   in case the file doesn't exists.
+ *   in case the file doesn't exists. DONE
  * - Extend that for other servers that are yet to have an
- *   existing Config file. (Format: <server_id>_config.json)
+ *   existing Config file. (Format: serverID_config.json) AND DONE (I'm so amazing 🥰)
  * @property {String} [guildID]
  * @property {Object} [settings]
  */
 
 
-class Config extends ConfigErrorHandler {
+class ConfigFile extends ConfigErrorHandler {
     
-      constructor(guildID) {
+      constructor(guildID = "") {
         super()
         if (guildID === "") {
-            throw new ConfigErrorHandler('The guildID cannot be empty.', 'InvalidGuildID')
+            throw new ConfigErrorHandler('The guildID is empty.', 'InvalidGuildID')
         }
         this.guildID = guildID
         try {
@@ -44,13 +44,13 @@ class Config extends ConfigErrorHandler {
       }
 
 
-      createMissingConfigFile(guildID) {
+    createMissingConfigFile(guildID = "") {
 
         const defaultSettings = JSON.parse(fs.readFileSync(`./configs/default.json`))
         fs.writeFileSync(`./configs/${guildID}_config.json`, JSON.stringify(defaultSettings))
         return defaultSettings
 
-      }
+    }
     //I don't really think this will have any actual use, considering
     //the Config Object gets remade every time.
 
@@ -66,20 +66,35 @@ class Config extends ConfigErrorHandler {
         return this.settings
     }
     
-    updateSetting(settingName, value) {
+    updateSetting(settingName = "", value = "") {
 
         this.settings = this.refreshConfigFile()
-        this.settings[settingName] = value
-        fs.writeFileSync(`./configs/${this.guildID}_config.json`, JSON.stringify(this.settings))
+        if (this.isSettingExist(settingName)) {
+            this.settings[settingName] = value
+            fs.writeFileSync(`./configs/${this.guildID}_config.json`, JSON.stringify(this.settings))
+            console.log(`The setting ${settingName} was accessed and updated to ${value}.`);
+        } else
+            throw new ConfigErrorHandler('The specified setting does not exist.', 'InvalidSettingName')
         
     }
 
-    getSetting(settingName) {
+    getSetting(settingName = "") {
 
         this.settings = this.refreshConfigFile()
         // console.log(this.settings[settingName])
-        return this.settings[settingName]
-    
+        if (this.isSettingExist(settingName)) {
+            console.log(`The setting ${settingName} was accessed.`);
+            return this.settings[settingName]
+        } else {
+            throw new ConfigErrorHandler(`The specified setting does not exist. Got ${settingName}`, 'InvalidSettingName')
+        }
+    }
+
+    isSettingExist(settingName = "") {
+        
+        this.settings = this.refreshConfigFile()
+        return this.settings[settingName] === "undefined"
+
     }
 
     toString() {
@@ -94,4 +109,4 @@ class Config extends ConfigErrorHandler {
     }
 }
 
-module.exports = Config;
+module.exports = ConfigFile;
